Started work on workout page
Created exercise card and content as components. Later content will be used to take the server data on exercises and convert content into a grid of exercise cards based on the exercises read from the server. Created a new page: WorkoutPage. This will be the page that shows the current workout the user is editing or displays past workouts. Currently all very rough.